Temple Interactable and Skills no longer happen if they would change nothing
Acquire Rectangle and Green from temple now.
With the addition of sound for each skill, it made sense to add some logic so that every time you pull up the HUD, it didn't activate all the skill since it would play a bunch of sounds even if nothing was visually changing. Saves some unnecessary processing to.