Fixed a movement bug and a color change bug
Movement bug: If you held a direction on one of the circles, you continue moving in whatever angle the circle pushes you towards.

Color bug: If you change color too close to a colored wall, you get trapped in your current color, but it thinks you are in the color you tried to switch to.