UI changes
Changed UI buttons to be our custom ones.
Added a button that returns the player to the main menu.
(To fully get it working we have to remove dont destroy on load from a few objects in the main scene. I didn't do this because I know people have some other changes).