Small tweaks to Aaron's puzzles
Just rearranging the tilemap a bit to make it easier to move and so you get stuck less. Also there where some tiles on the wrong layer that I fixed.

Fixed a few softlocks that I was able to find.

Fixed a bug I found with the pressure plates where if the player is on the pressure plate with an ice block and then moves away it would break.

Moved everything to fit on the grid a bit better.

Moved the sphere upgrade to be at the end of Aaron's puzzles.