NPC dialogue, king vision width, and fixed bug
Added periods to a lot of the NPC dialogue. Also corrected some grammar and other punctuation.

Made the king's vision wider so that you have to talk to him before getting to the end.

One of the ice pushables in the pressure plate puzzle was locked on both the x and y axis for movement. I unlocked its x axis so you can beat the puzzle now.