Grid based movement.
Player now walks on the grid.
New versions of the ice and pushable are being worked on in the new scene, they are mostly done.
Changing shapes was updated to still work.
Rotating shapes was updated to skill work.

Make the change to rotating the shape and such look better.